maple feedback

+++ jeditrader/axes.h
@@ -4,7 +4,7 @@
 #include "window.h"
 
 Axes axes_default();
-void axes_init(Axes *axes, double aspect_ratio);
+void axes_init(Axes *axes, float aspect_ratio);
 void axes_update(Axes *axes, Window* window);
 void axes_render_frame(Axes* axes, mat4 g_world);
